1039_lld long long int

+++ main.cpp
@@ -1,9 +1,9 @@
 #include <iostream>
 
 int main(){
-  long int a, b;
-  scanf("%ld %ld",&a, &b);
-  printf("%ld",a + b);
+  long long int a, b;
+  scanf("%lld %lld",&a, &b);
+  printf("%lld",a + b);
   }
 
 //아스키 코드값이 정수값으로 출력하면 정수값이 나옴
